API Documentation

If you're a newcomer to Legolas.jl, please familiarize yourself with via the tour before diving into this documentation.

Legolas Schemas and Rows

Legolas.@row — Macro

TODO

Legolas.Row — Type

TODO

Legolas.Schema — Type

TODO

Legolas.is_valid_schema_name — Function

TODO

Legolas.schema_name — Function

TODO

Legolas.schema_version — Function

TODO

Legolas.schema_qualified_string — Function

TODO

Legolas.schema_parent — Function

TODO

Legolas.transform — Function

TODO

Validating/Writing/Reading Legolas Tables

Legolas.validate — Function

TODO

Legolas.write — Function

TODO

Legolas.read — Function

TODO

Utilities

Legolas.lift — Function

lift(f, x)

Return f(x) unless x isa Union{Nothing,Missing}, in which case return missing.

This is particularly useful when handling values from Arrow.Table, whose null values may present as either missing or nothing depending on how the table itself was originally constructed.

Legolas.assign_to_table_metadata! — Function

TODO

Note that we intend to eventually migrate this function from Legolas.jl to a more appropriate package.

Legolas.gather — Function

gather(column_name, tables...; extract=((table, idxs) -> view(table, idxs, :)))

Gather rows from tables into a unified cross-table index along column_name. Returns a Dict whose keys are the unique values of column_name across tables, and whose values are tuples of the form:

(rows_matching_key_in_table_1, rows_matching_key_in_table_2, ...)

The provided extract function is used to extract rows from each table; it takes as input a table and a Vector{Int} of row indices, and returns the corresponding subtable. The default definition is sufficient for DataFrames tables.

Note that this function may internally call Tables.columns on each input table, so it may be slower and/or require more memory if any(!Tables.columnaccess, tables).

Note that we intend to eventually migrate this function from Legolas.jl to a more appropriate package.

Legolas.materialize — Function

materialize(table)

Return a fully deserialized copy of table.

This function is useful when table has built-in deserialize-on-access or conversion-on-access behavior (like Arrow.Table) and you'd like to pay such access costs upfront before repeatedly accessing the table. For example:

TODO: make this example runnable

julia> items = read_table(path_to_file);

# iterate through all elements of `items.nested_structures`
julia> @time foreach(identity, (nested_structure for nested_structure in items.nested_structures));
0.000126 seconds (306 allocations: 6.688 KiB)

julia> materialized = Onda.materialize(items);

julia> @time foreach(identity, (nested_structure for nested_structure in materialized.nested_structures));
  0.000014 seconds (2 allocations: 80 bytes)

Note that we intend to eventually migrate this function from Legolas.jl to a more appropriate package.