API Documentation

Note

If you're a newcomer to Legolas.jl, please familiarize yourself with the tour before diving into this documentation.

Legolas Schemas

Legolas.Schema — Type

Legolas.Schema{name,version}

A type representing a Legolas schema. The name (a Symbol) and version (an Integer) are surfaced as type parameters, allowing them to be utilized for dispatch.

For more details and examples, please see Legolas.jl/examples/tour.jl and the "Schema-Related Concepts/Conventions" section of the Legolas.jl documentation.

The constructor Schema{name,version}() will throw an ArgumentError if version is not non-negative.

See also: Legolas.@schema

Legolas.@schema — Macro

@schema("name@version", field_expressions...)
@schema("name@version > parent_name@parent_version", field_expressions...)

Declare a new Legolas.Schema{name,version} and overload relevant Legolas method definitions with schema-specific behaviors in accordance with the declared schema identifier, parent identifier (if present), and required fields.

Each element of field_expressions defines a required field for the declared schema, and is an expression of the form field::F = rhs where:

field is the corresponding field's name

::F denotes the field's type constraint (if elided, defaults to ::Any).

rhs is the expression which produces field::F (if elided, defaults to field).

Accounting for all of the aforementioned allowed elisions, possible valid forms of elements of field_expressions include:

field::F = rhs

field::F (interpreted as field::F = field)

field = rhs (interpreted as field::Any = rhs)

field (interpreted as field::Any = field)

Note that:

A Legolas.SchemaDeclarationError will be thrown if:

The provided schema identifier is invalid or is not a string literal

If there are no required field expressions, duplicate required fields are declared, or a given expression is invalid

(if a parent schema is specified) This schema declaration does not comply with its parent's schema declaration, or the parent hasn't yet been declared

This macro expects to be evaluated within top-level scope.

For more details and examples, please see Legolas.jl/examples/tour.jl and the "Schema-Related Concepts/Conventions" section of the Legolas.jl documentation.

Legolas.is_valid_schema_name — Function

Legolas.is_valid_schema_name(x::AbstractString)

Return true if x is a valid schema name, return false otherwise.

Valid schema names are lowercase, alphanumeric, and may contain hyphens or periods.

Legolas.parse_schema_identifier — Function

Legolas.parse_schema_identifier(id::AbstractString)

Given a valid schema identifier id of the form:

$(names[1])@$(versions[1]) > $(names[2])@$(versions[2]) > ... > $(names[n])@$(versions[n])

return an n element Vector{Schema} whose ith element is Schema(names[i], versions[i]).

Throws an ArgumentError if the provided string is not a valid schema identifier.

For details regarding valid schema identifiers and their structure, see the "Schema-Related Concepts/Conventions" section of the Legolas.jl documentation.

Legolas.schema_name — Function

Legolas.schema_name(::Legolas.Schema{name})

Return name.

Legolas.schema_version — Function

Legolas.schema_version(::Legolas.Schema{name,version})

Return version.

Legolas.schema_identifier — Function

Legolas.schema_identifier(::Legolas.Schema)

Return this Legolas.Schema's fully qualified schema identifier. This string is serialized as the "legolas_schema_qualified" field value in table metadata for table written via Legolas.write.

Legolas.schema_parent — Function

Legolas.schema_parent(::Legolas.Schema)

Return the Legolas.Schema instance that corresponds to the parent of the given Legolas.Schema.

Legolas.schema_fields — Function

Legolas.schema_fields(schema::Legolas.Schema)

Return a NamedTuple{...,Tuple{Vararg{DataType}} whose fields take the form:

<name of field required by schema> = <field's type>

If schema has a parent, the returned fields will include schema_fields(schema_parent(schema)).

Legolas.schema_declaration — Function

Legolas.schema_declaration(schema::Legolas.Schema)

Return a Pair{String,Dict{Symbol,Expr}} of the form

schema_identifier_used_for_declaration::String => required_field_expressions::Dict{Symbol,Expr}

where each key-value pair in required_field_expressions takes the form:

required_field_name::Symbol => normalized_required_field_assignment_statement::Expr

Note that schema_declaration is primarily intended to be used for interactive discovery purposes, and does not include the contents of schema_declaration(schema_parent(schema)).

Legolas.schema_declared — Function

Legolas.schema_declared(schema::Legolas.Schema{name,version})

Return true if the schema name@version has been declared via @schema in the current Julia session; return false otherwise.

Legolas.row — Function

Legolas.row(schema::Schema; fields...)
Legolas.row(schema::Schema, fields)

Return a Tables.AbstractRow-compliant value whose contents are provided by fields and validated/transformed in accordance with schema. If the non-keyword-arguments-based method is used, fields must be a Tables.AbstractRow-compliant value.

For more details and examples, please see Legolas.jl/examples/tour.jl.

Validating/Writing/Reading Legolas Tables

Legolas.extract_legolas_schema — Function

Legolas.extract_legolas_schema(table)

Attempt to extract Arrow metadata from table via Arrow.getmetadata(table).

If Arrow metadata is present and contains "legolas_schema_qualified" => s, return first(parse_schema_identifier(s))

Otherwise, return nothing.

Legolas.write — Function

Legolas.write(io_or_path, table, schema::Schema; validate::Bool=true, kwargs...)

Write table to io_or_path, inserting the appropriate legolas_schema_qualified field in the written out Arrow metadata.

If validate is true, Legolas.validate will be called on the table before it written out.

Any other provided kwargs are forwarded to an internal invocation of Arrow.write.

Note that io_or_path may be any type that supports Base.write(io_or_path, bytes::Vector{UInt8}).

Legolas.read — Function

Legolas.read(io_or_path; validate::Bool=true)

Read and return an Arrow.Table from io_or_path.

If validate is true, Legolas.read will attempt to extract a Legolas.Schema from the deserialized Arrow.Table's metadata and use Legolas.validate to verify that the table's Table.Schema complies with the extracted Legolas.Schema before returning the table.

Note that io_or_path may be any type that supports Base.read(io_or_path)::Vector{UInt8}.

Utilities

Legolas.lift — Function

lift(f, x)

Return f(x) unless x isa Union{Nothing,Missing}, in which case return missing.

This is particularly useful when handling values from Arrow.Table, whose null values may present as either missing or nothing depending on how the table itself was originally constructed.

See also: construct

lift(f)

Returns a curried function, x -> lift(f,x)

Legolas.construct — Function

construct(T::Type, x)

Construct T(x) unless x is of type T, in which case return x itself. Useful in conjunction with the lift function for types which don't have a constructor which accepts instances of itself (e.g. T(::T)).

Examples

julia> using Legolas: construct

julia> construct(Float64, 1)
1.0

julia> Some(Some(1))
Some(Some(1))

julia> construct(Some, Some(1))
Some(1)

Use the curried form when using lift:

julia> using Legolas: lift, construct

julia> lift(Some, Some(1))
Some(Some(1))

julia> lift(construct(Some), Some(1))
Some(1)

Legolas.guess_schema — Function

Legolas.guess_schema(table)

Returns a Table.Schema matching the provided table if such a Tables.Schema can be determined, otherwise return nothing.

This function is similar to Tables.schema(table), but may undertake additional work to more aggressively attempt to determine a matching Table.Schema, even when it might not be directly evident from typeof(table).

Specifically, this function may call Tables.columns(table) as a preprocessing step.

Missing docstring.

Missing docstring for Legolas.assign_to_table_metadata!. Check Documenter's build log for details.

Legolas.gather — Function

Legolas.gather(column_name, tables...; extract=((table, idxs) -> view(table, idxs, :)))

Gather rows from tables into a unified cross-table index along column_name. Returns a Dict whose keys are the unique values of column_name across tables, and whose values are tuples of the form:

(rows_matching_key_in_table_1, rows_matching_key_in_table_2, ...)

The provided extract function is used to extract rows from each table; it takes as input a table and a Vector{Int} of row indices, and returns the corresponding subtable. The default definition is sufficient for DataFrames tables.

Note that this function may internally call Tables.columns on each input table, so it may be slower and/or require more memory if any(!Tables.columnaccess, tables).

Note that we intend to eventually migrate this function from Legolas.jl to a more appropriate package.

Legolas.locations — Function

locations(collections::Tuple)

Return a Dict whose keys are the set of all elements across all provided collections, and whose values are the indices that locate each corresponding element across all provided collecitons.

Specifically, locations(collections)[k][i] will return a Vector{Int} whose elements are the index locations of k in collections[i]. If !(k in collections[i]), this Vector{Int} will be empty.

For example:

julia> Legolas.locations((['a', 'b', 'c', 'f', 'b'],
                          ['d', 'c', 'e', 'b'],
                          ['f', 'a', 'f']))
Dict{Char, Tuple{Vector{Int64}, Vector{Int64}, Vector{Int64}}} with 6 entries:
  'f' => ([4], [], [1, 3])
  'a' => ([1], [], [2])
  'c' => ([3], [2], [])
  'd' => ([], [1], [])
  'e' => ([], [3], [])
  'b' => ([2, 5], [4], [])

This function is useful as a building block for higher-level tabular operations that require indexing/grouping along specific sets of elements.

Legolas.materialize — Function

Legolas.materialize(table)

Return a fully deserialized copy of table.

This function is useful when table has built-in deserialize-on-access or conversion-on-access behavior (like Arrow.Table) and you'd like to pay such access costs upfront before repeatedly accessing the table.

Note that we intend to eventually migrate this function from Legolas.jl to a more appropriate package.